feat: LibraryRepository 생성 및 DI 모듈 등록
