+++ data/library/src/main/java/com/into/websoso/data/library/repository/MyLibraryRepository.kt
@@ -20,10 +20,8 @@ import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.flatMapLatest
 import kotlinx.coroutines.flow.map
 import javax.inject.Inject
-import javax.inject.Singleton
 
-@Singleton
-class MyLibraryRepository
+internal class MyLibraryRepository
     @Inject
     constructor(
         filterRepository: FilterRepository,
